Extract AI theme builder into optional :ui:theming:ai module
Moves the on-device AI theme generation feature (Halogen + Gemini Nano
via ML Kit) out of :ui:theming:impl and :ui:theming:ui into a new
optional :ui:theming:ai module, so proprietary Google dependencies can
be excluded from a future foss product flavor (#911).

- New AiThemeBuilder seam in :ui:theming:api (isSupported +
  observeIsAvailable) with a NoOpAiThemeBuilder reached via
  kotlin-inject default-argument fallback when the module is absent
- HalogenAiThemeBuilder contributed to AppScope from the ai module
- ThemePickerPresenter routes existing AI themes to the regular theme
  builder when the AI module isn't in the build
- ML Kit GenAI, halogen (core/compose/engine/provider-nano), cadence,
  and androidx.shapes all move out of the always-included modules
- Wired via standard/alpha/betaImplementation in :app:android; iOS and
  desktop drop the module (the Nano provider was always null there)
