+++ ui/theming/ai/src/androidMain/kotlin/app/campfire/ui/theming/ai/halogen/HalogenProviderFactory.android.kt
@@ -1,0 +1,5 @@
+package app.campfire.ui.theming.ai.halogen
+
+import halogen.HalogenLlmProvider
+
+internal actual fun createNanoProvider(): HalogenLlmProvider? = DiagnosticGeminiNanoProvider()
