+++ ui/theming/ai/src/commonMain/composeResources/values/ui_theming_ai_strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <string name="action_back">Back</string>
+</resources>
