+++ ui/theming/ai/src/commonMain/kotlin/app/campfire/ui/theming/ai/halogen/HalogenProviderFactory.kt
@@ -1,0 +1,12 @@
+package app.campfire.ui.theming.ai.halogen
+
+import halogen.HalogenLlmProvider
+
+/**
+ * Build the platform's on-device LLM provider for Halogen, or `null` if the platform
+ * doesn't support on-device GenAI.
+ *
+ * - **Android**: returns [DiagnosticGeminiNanoProvider] (Gemini Nano via ML Kit).
+ * - **iOS / Desktop**: returns `null` — the AI Theme Builder feature is hidden.
+ */
+internal expect fun createNanoProvider(): HalogenLlmProvider?
