+++ ui/theming/ai/src/commonMain/kotlin/app/campfire/ui/theming/ai/ui/emptystate/components/ParticleEmission.kt
@@ -1,0 +1,16 @@
+package app.campfire.ui.theming.ai.ui.emptystate.components
+
+import androidx.compose.ui.geometry.Offset
+import com.r0adkll.cadence.game.ecs.Component
+
+/**
+ * This is attached to an entity to give it emission information
+ * such as its exit velocity and acceleration. This component
+ * allows us to act upon the entity in a particular system or not.
+ */
+class ParticleEmission(
+  val exitVelocity: Offset,
+  val exitAcceleration: Offset,
+  val initialAngularVelocity: Float,
+  val exitAngularVelocity: Float,
+) : Component
