+++ ui/theming/ai/src/jvmMain/kotlin/app/campfire/ui/theming/ai/halogen/HalogenProviderFactory.jvm.kt
@@ -1,0 +1,5 @@
+package app.campfire.ui.theming.ai.halogen
+
+import halogen.HalogenLlmProvider
+
+internal actual fun createNanoProvider(): HalogenLlmProvider? = null
