+++ ui/theming/impl/build.gradle.kts
@@ -32,10 +32,6 @@ kotlin {
         implementation(libs.compose.runtime)
         implementation(libs.compose.ui)
 
-        implementation(libs.halogen.core)
-        implementation(libs.halogen.compose)
-        implementation(libs.halogen.engine)
-
         implementation(libs.stately.concurrent.collections)
         implementation(libs.sqldelight.coroutines)
         implementation(libs.sqldelight.async)
@@ -46,10 +42,6 @@ kotlin {
     androidMain {
       dependencies {
         implementation(libs.sqldelight.android)
-        implementation(libs.halogen.provider.nano)
-        // Direct ML Kit GenAI dependency so DiagnosticGeminiNanoProvider can call
-        // checkStatus() and surface the raw FeatureStatus to logcat.
-        implementation("com.google.mlkit:genai-prompt:1.0.0-beta4")
       }
     }
