--- ui/theming/impl/src/androidMain/kotlin/app/campfire/ui/theming/halogen/HalogenProviderFactory.android.kt
@@ -1,5 +1,0 @@
-package app.campfire.ui.theming.halogen
-
-import halogen.HalogenLlmProvider
-
-internal actual fun createNanoProvider(): HalogenLlmProvider? = DiagnosticGeminiNanoProvider()
