--- ui/theming/impl/src/iosMain/kotlin/app/campfire/ui/theming/halogen/HalogenProviderFactory.ios.kt
@@ -1,5 +1,0 @@
-package app.campfire.ui.theming.halogen
-
-import halogen.HalogenLlmProvider
-
-internal actual fun createNanoProvider(): HalogenLlmProvider? = null
