--- ui/theming/impl/src/jvmMain/kotlin/app/campfire/ui/theming/halogen/HalogenProviderFactory.jvm.kt
@@ -1,5 +1,0 @@
-package app.campfire.ui.theming.halogen
-
-import halogen.HalogenLlmProvider
-
-internal actual fun createNanoProvider(): HalogenLlmProvider? = null
