+++ ui/theming/ui/build.gradle.kts
@@ -13,8 +13,6 @@ kotlin {
 
         implementation(libs.color.picker.compose)
         implementation(libs.circuitx.overlays)
-        implementation(libs.cadence)
-        implementation(libs.androidx.shapes)
       }
     }
   }
