--- ui/theming/ui/src/commonMain/kotlin/app/campfire/ui/theming/ui/ai/emptystate/EmitArea.kt
@@ -1,44 +1,0 @@
-package app.campfire.ui.theming.ui.ai.emptystate
-
-import androidx.compose.ui.unit.IntOffset
-import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.unit.toSize
-import androidx.compose.ui.util.fastRoundToInt
-import kotlin.concurrent.atomics.AtomicInt
-import kotlin.concurrent.atomics.ExperimentalAtomicApi
-import kotlin.concurrent.atomics.fetchAndIncrement
-import kotlin.random.Random
-
-sealed class EmitArea(val size: IntSize) {
-  abstract fun randomPosition(): IntOffset
-
-  @OptIn(ExperimentalAtomicApi::class)
-  class Distributed(
-    size: IntSize,
-    minDistance: Float,
-    maxParticles: Int,
-  ) : EmitArea(size) {
-    private val pointIndex = AtomicInt(0)
-    private val points = generatePoissonDiskSamples(
-      size = size.toSize(),
-      minDistance = minDistance.toDouble(),
-      k = maxParticles,
-    )
-
-    override fun randomPosition(): IntOffset {
-      val index = pointIndex.fetchAndIncrement() % points.size
-      return points[index].let {
-        IntOffset(it.x.fastRoundToInt(), it.y.fastRoundToInt())
-      }
-    }
-  }
-
-  class Bottom(size: IntSize) : EmitArea(size) {
-    override fun randomPosition(): IntOffset {
-      return IntOffset(
-        x = Random.nextInt(size.width),
-        y = size.height,
-      )
-    }
-  }
-}
