+++ ui/theming/ui/src/commonMain/kotlin/app/campfire/ui/theming/ui/builder/composables/IconPicker.kt
@@ -29,7 +29,7 @@ private val GridPadding = 8.dp
 private val GridWidth = IconSize * ColumnCount + IconSpacing * (ColumnCount - 1) + GridPadding * 2
 
 @Composable
-internal fun IconPicker(
+fun IconPicker(
   icon: AppTheme.Icon,
   onIconClick: (AppTheme.Icon) -> Unit,
   modifier: Modifier = Modifier,
